// src/app/core/movimentacoes-contabeis-dto.ts
export class MovimentacoesContabeisDto {
	id: number;
	data: Date;
	valor: number;
	idPrograma01: number;
	nomePrograma01: string;
	idProjeto01: number;
	nomeProjeto01: string;
	idCategoriaOrigem01: number;
	planoContaOrigem01: string;
	codigoPlanoContaOrigem01: string;
	idCategoriaDestino01: number;
	planoContaDestino01: string;
	codigoPlanoContaDestino01: string;
	idPrograma02: number;
	nomePrograma02: string;
	idProjeto02: number;
	nomeProjeto02: string;
	idCategoriaOrigem02: number;
	planoContaOrigem02: string;
	codigoPlanoContaOrigem02: string;
	idCategoriaDestino02: number;
	planoContaDestino02: string;
	codigoPlanoContaDestino02: string;

}